Power control interface update
