+++ libs/power/Include/power/power.h
@@ -3,8 +3,6 @@
 
 #include "system.h"
 
-EXTERNC_BEGIN
-
 namespace services
 {
     namespace power
@@ -28,12 +26,35 @@ namespace services
              * @brief Performs sail opening sequence
              */
             virtual void OpenSail() = 0;
+
+            /**
+             * @brief Sets state of main thermal knife LCL
+             * @param enabled true to turn on LCL, false to turn off
+             * @return Operation status
+             */
+            virtual bool MainThermalKnife(bool enabled) = 0;
+            /**
+             * @brief Sets state of redundant thermal knife LCL
+             * @param enabled true to turn on LCL, false to turn off
+             * @return Operation status
+             */
+            virtual bool RedundantThermalKnife(bool enabled) = 0;
+
+            /**
+             * @brief Enables main sail burn switch
+             * @return Operation status
+             */
+            virtual bool EnableMainSailBurnSwitch() = 0;
+
+            /**
+             * @brief Enables redundant sail burn switch
+             * @return Operation status
+             */
+            virtual bool EnableRedundantSailBurnSwitch() = 0;
         };
 
         /** @} */
     }
 }
-
-EXTERNC_END
 
 #endif /* POWER_H_ */
